+++ src/host-install-service-systemd.sh
@@ -12,11 +12,13 @@ echo "Service file for $APP_NAME was created as $serviceFile"
 
 if command su --help &>/dev/null ; then
 	echo "Trying to install it now as root user..."
-	su -c "cp -aT '$serviceFile' '$unitFile' && systemctl daemon-reload && (systemctl restart ${APP_NAME}.service || (systemctl start ${APP_NAME}.service && systemctl enable ${APP_NAME}.service))"\
-		&& rm $serviceFile\
-		&& echo "Done"\
-		&& exit 0
-	echo "Failed"
+	([ ! -f "$unitFile" ] && exit 0 || (read -r -p "'$unitFile' exists. Overwrite? [n/Y] " confirmation && [ "$confirmation" = 'Y' ] || exit 1))\
+		&& (su -c "cp -aT '$serviceFile' '$unitFile' && systemctl daemon-reload && (systemctl restart ${APP_NAME}.service || (systemctl start ${APP_NAME}.service && systemctl enable ${APP_NAME}.service))"\
+			&& rm $serviceFile\
+			&& echo "Done"\
+			|| echo "Failed" && exit 1)\
+		&& exit 0\
+		|| echo "Cancelled"
 	echo ""
 fi
